Crashes CP when run with 64-bit Scipy
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CellProfiler@9625 1fc53939-2000-0410-845c-e8453a809027

+++ cellprofiler/modules/measureimagequality.py
@@ -440,7 +440,7 @@ class MeasureImageQuality(cpm.CPModule):
         grid = i * m + j + 1
         if image.has_mask:
             grid[np.logical_not(image.mask)] = 0
-        grid_range = np.arange(0, m*n+1)
+        grid_range = np.arange(0, m*n+1,dtype=int)
         #
         # Do the math per label
         #
